The code and their headers
this is the code that controls the program. they connect to the ui's and calculations, allowing for the code to run and for the buttons and other functions to work. there is also the .txt file and .qrc file along with the image for the .qrc file which enables the border image for the window